requirements.txt
google-genai
python-dotenv
fastapi==0.135.1
uvicorn[standard]==0.41.0
